Update to bigint for mpan, mprn

+++ db/migrate/20180628091550_update_meter_attributes.rb
@@ -1,6 +1,6 @@
 class UpdateMeterAttributes < ActiveRecord::Migration[5.2]
   def change
-    add_column :meters, :mpan_mprn, :integer
+    add_column :meters, :mpan_mprn, :bigint
     add_column :meters, :meter_serial_number, :text
     add_column :meters, :solar_pv,  :boolean, default: false
     add_column :meters, :storage_heaters, :boolean, default: false
